Early version with some terrible piece matching, single winnowing
Couple early versions, where I was trying to find a way to do wildcards that bounced front to back, winnowing down. This would have reduced maximum memory size, but when I started into the winnowing operations it was getting hard.

I'd also wanted an optimization that went fragment by fragment, rather than char by char, but I was still having to step through each fragment most of the time, so it it was again just more painful to manage state.